employee_list_api/requirements.txt
Django==2.1.4
django-crispy-forms==1.7.2
djangorestframework==3.9.1
Pillow==5.3.0
psycopg2-binary==2.7.6.1
pytz==2018.7
